navbar completed
