+++ src/Components/Navbar/Navbar.css
@@ -1,7 +1,16 @@
+.navbar-logo{
+width: 130px;
+
+}
 .navbar{
     display: flex;
-    /* justify-content: flex-end; */
-    justify-content: space-evenly;
+   
+    /* justify-content: space-evenly; */
+   
+}
+.navbar-main{
+    display: flex;
+    justify-content: space-between;
 }
 
 .navbar-title {
@@ -12,6 +21,7 @@
   line-height: 27px;
   color: #000000;
   text-decoration: none;
+  padding-left: 20px;
 }
 .navbar-title:hover{
     color:orange;
@@ -19,4 +29,8 @@
 .navbar-button{
     margin-top: 10px;
     margin-left: 20px;
+}
+ul{
+    display: flex;
+    list-style: none;
 }
