+++ src/Pages/Home/Home.js
@@ -1,0 +1,12 @@
+import React from "react";
+import Navbar from "../../Components/Navbar/Navbar";
+
+const Home = () => {
+  return (
+    <div>
+      <Navbar />
+    </div>
+  );
+};
+
+export default Home;
